refactor(utils): split file_logging into thin facade + sibling modules
Break the 839-line `file_logging.rs` module into focused siblings using
the `#[path]`-included submodule pattern established by the watcher and
action-chain refactors:

- `file_logging_config.rs` — `RotationPolicy`, `FileLoggingConfig`,
  `FileLoggingError`, env-var parsing.
- `file_logging_writer.rs` — `RollingFileWriter`, `Inner` rotation
  state, `CalendarDate`, and the filesystem helpers (`open_append`,
  `current_path`, `rotated_path`, `prune_old`).
- `file_logging_python.rs` — PyO3 wrappers (`PyFileLoggingConfig`,
  `py_init_file_logging`, `py_shutdown_file_logging`, `py_flush_logs`,
  `_default_file_logging_settings`).
- `file_logging_tests.rs` — rotation, retention, and install/shutdown
  idempotency tests moved out of the main file.

`file_logging.rs` is now a thin facade that owns only the public
install / shutdown / `flush_logs` entry points and the process-wide
`FileLoggingHandles` + `HANDLES` slot. No public API changes — all
re-exports flow through `pub use`.

Documentation: `docs/guide/architecture.md` now describes the new
layout under the `dcc-mcp-utils` section.
